Created check_address()
check_address() checks if a Bitcoin Cash address is valid.
It also prepends the 'bitcoincash:' prefix if the user has not included it.
It takes update and address as arguments and returns the BCH address if it is correct, False if not.

This commit does not include tests for this function.